+++ src/connectors/range/connectRange.ts
@@ -476,7 +476,14 @@ const connectRange: ConnectRange = function connectRange(
 
         const [lowerBound, upperBound] = value.split(':').map(parseFloat);
 
-        if (isFiniteNumber(lowerBound)) {
+        if (
+          isFiniteNumber(lowerBound) &&
+          (!isFiniteNumber(minBound) || minBound < lowerBound)
+        ) {
+          widgetSearchParameters = widgetSearchParameters.removeNumericRefinement(
+            attribute,
+            '>='
+          );
           widgetSearchParameters = widgetSearchParameters.addNumericRefinement(
             attribute,
             '>=',
@@ -484,7 +491,14 @@ const connectRange: ConnectRange = function connectRange(
           );
         }
 
-        if (isFiniteNumber(upperBound)) {
+        if (
+          isFiniteNumber(upperBound) &&
+          (!isFiniteNumber(maxBound) || upperBound < maxBound)
+        ) {
+          widgetSearchParameters = widgetSearchParameters.removeNumericRefinement(
+            attribute,
+            '<='
+          );
           widgetSearchParameters = widgetSearchParameters.addNumericRefinement(
             attribute,
             '<=',